+++ src/Tracker.cpp
@@ -1,0 +1,17 @@
+/**
+ * @file   Tracker.cpp
+ * @brief  Implementations of tracker class in SLAM system.
+ * @author Charlie Li
+ * @date   2019.08.09
+ */
+
+#include "Tracker.hpp"
+
+namespace SLAM_demo {
+
+Tracker::Tracker(System::Mode eMode):
+    meMode(eMode)
+{
+}
+
+} // namespace SLAM_demo