+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Latest version
 - Add system initialzation scheme (5%)
+  - Add basic structure of Tracker class (10%)
 - Add system parameter loading scheme (done)
   - System parameter yaml file (done)
   - Create a singleton Config class for global cfgs (done)